moved sidebar into gagestats
