+++ src/app/gagestats/gs-sidebar/gs-sidebar.component.scss
