+++ src/app/gagestats/gs-sidebar/gs-sidebar.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { NSSService } from '../shared/services/app.service';
-import { Agency } from '../shared/interfaces/agency';
-import { Stationtype } from '../shared/interfaces/stationtype';
-import { IMultiSelectSettings, IMultiSelectTexts} from '../../../node_modules/angular-2-dropdown-multiselect';
+import { NSSService } from 'app/shared/services/app.service';
+import { Agency } from 'app/shared/interfaces/agency';
+import { Stationtype } from 'app/shared/interfaces/stationtype';
+import { IMultiSelectSettings, IMultiSelectTexts} from '../../../../node_modules/angular-2-dropdown-multiselect';
 
 
 @Component({
